+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -667,6 +667,36 @@ public class CollUtil {
 		return currentAlaDatas;
 	}
 
+	/**
+	 * 是否至少有一个符合判断条件
+	 *
+	 * @param <T> 集合元素类型
+	 * @param collection 集合
+	 * @param predicate 自定义判断函数
+	 * @return 是否有一个值匹配 布尔值
+	 */
+	public static <T>boolean anyMatch(Collection<T> collection,Predicate<T> predicate){
+		if(isEmpty(collection)){
+			return Boolean.FALSE;
+		}
+		return collection.stream().anyMatch(predicate);
+	}
+
+	/**
+	 * 是否全部匹配判断条件
+	 *
+	 * @param <T> 集合元素类型
+	 * @param collection 集合
+	 * @param predicate  自定义判断函数
+	 * @return 是否全部匹配 布尔值
+	 */
+	public static <T>boolean allMatch(Collection<T> collection,Predicate<T> predicate){
+		if(isEmpty(collection)){
+			return Boolean.FALSE;
+		}
+		return collection.stream().allMatch(predicate);
+	}
+
 	// ----------------------------------------------------------------------------------------------- new HashSet
 
 	/**
@@ -2145,7 +2175,13 @@ public class CollUtil {
 		if (value instanceof Iterator) {
 			iter = (Iterator) value;
 		} else if (value instanceof Iterable) {
-			iter = ((Iterable) value).iterator();
+			if(value instanceof Map && BeanUtil.isBean(TypeUtil.getClass(elementType))){
+				//https://github.com/dromara/hutool/issues/3139
+				// 如果值为Map，而目标为一个Bean，则Map应整体转换为Bean，而非拆分成Entry转换
+				iter = new ArrayIter<>(new Object[]{value});
+			}else{
+				iter = ((Iterable) value).iterator();
+			}
 		} else if (value instanceof Enumeration) {
 			iter = new EnumerationIter<>((Enumeration) value);
 		} else if (ArrayUtil.isArray(value)) {