+++ hutool-core/src/main/java/cn/hutool/core/io/FileUtil.java
@@ -3459,18 +3459,7 @@ public class FileUtil extends PathUtil {
 	 */
 	public static File checkSlip(File parentFile, File file) throws IllegalArgumentException {
 		if (null != parentFile && null != file) {
-			String parentCanonicalPath;
-			String canonicalPath;
-			try {
-				parentCanonicalPath = parentFile.getCanonicalPath();
-				canonicalPath = file.getCanonicalPath();
-			} catch (IOException e) {
-				// issue#I4CWMO@Gitee
-				// getCanonicalPath有时会抛出奇怪的IO异常，此时忽略异常，使用AbsolutePath判断。
-				parentCanonicalPath = parentFile.getAbsolutePath();
-				canonicalPath = file.getAbsolutePath();
-			}
-			if (false == canonicalPath.startsWith(parentCanonicalPath)) {
+			if (false == isSub(parentFile, file)) {
 				throw new IllegalArgumentException("New file is outside of the parent dir: " + file.getName());
 			}
 		}